# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmPay - API de Gestion des Comptes Bancaires",
        "description": "API OmPay pour la gestion des comptes bancaires avec création automatique d'utilisateurs et clients",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://appli-laravel-groupe.onrender.com/",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes",
                "description": "Récupère la liste paginée des comptes avec possibilité de filtrage et tri",
                "operationId": "getComptes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "inactif",
                                "bloqué"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par titulaire ou numéro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dateCreation",
                                "solde",
                                "titulaire"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte",
                "description": "Crée un nouveau compte bancaire. Si le client n'existe pas, il est créé automatiquement avec génération de mot de passe et code. Un email et un SMS sont envoyés automatiquement.",
                "operationId": "createCompte",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCompteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Pagination": {
                "title": "Pagination",
                "description": "Informations de pagination",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "Links": {
                "title": "Links",
                "description": "Liens de navigation",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://api.banque.example.com/api/v1/comptes?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://api.banque.example.com/api/v1/comptes?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://api.banque.example.com/api/v1/comptes?page=2"
                    }
                },
                "type": "object"
            },
            "Compte": {
                "title": "Compte",
                "description": "Objet représentant un compte bancaire",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "C-20251025-ABCD"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Mamadou Diallo"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "devise": {
                        "type": "string",
                        "enum": [
                            "FCFA",
                            "XOF",
                            "EUR",
                            "USD"
                        ],
                        "example": "FCFA"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-25T17:33:20Z"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "inactif",
                            "bloqué"
                        ],
                        "example": "actif"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "informationsBlocage": {
                        "description": "Informations de blocage pour les comptes bloqués",
                        "properties": {
                            "dateDebutBlocage": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-10-30T12:00:00Z"
                            },
                            "dateFinBlocage": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-11-02T12:00:00Z"
                            },
                            "dureeBlocageJours": {
                                "description": "Durée du blocage en jours",
                                "type": "integer",
                                "example": 3
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-10-25T17:33:20Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ClientData": {
                "required": [
                    "prenom",
                    "nom",
                    "email",
                    "telephone"
                ],
                "properties": {
                    "prenom": {
                        "description": "Prénom du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "nom": {
                        "description": "Nom du client",
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "description": "Adresse email",
                        "type": "string",
                        "format": "email"
                    },
                    "telephone": {
                        "description": "Numéro de téléphone",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "Adresse",
                        "type": "string",
                        "maxLength": 255
                    },
                    "nci": {
                        "description": "Numéro CNI",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateCompteRequest": {
                "required": [
                    "solde",
                    "client"
                ],
                "properties": {
                    "solde": {
                        "description": "Solde initial (minimum 10 000 FCFA)",
                        "type": "number",
                        "minimum": 10000
                    },
                    "client": {
                        "$ref": "#/components/schemas/ClientData"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Message d'erreur",
                        "type": "string"
                    },
                    "errors": {
                        "description": "Détails des erreurs de validation",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Message de succès",
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Compte"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token JWT ici",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Gestion des comptes bancaires"
        }
    ]
}